Working Lex Parser
Getting this to compile was a pain. But It works now.
A template for defining a language for the input file exists now.
We just need to write a regex for the different types of lines, there
should only be four or five. The MTS file exists now, not much in it
except for the heap template. Now I need to get the MTS to point to a
specific file.